Add files via upload
Since the seminal paper by Dickey and Fuller in 1979, unit-root tests have conditioned the standard approaches to analysing time series with strong serial dependence in mean behaviour, the focus being placed on the detection of eventual unit roots in an autoregressive model fitted to the series. The Author propose a completely different method to test for the type of long-wave patterns observed not only in unit-root time series but also in series following more complex data-generating mechanisms. To this end, his testing device analyses the unit-root persistence exhibited by the data while imposing very few constraints on the generating mechanism. We call our device the range unit-root (RUR) test since it is constructed from the running ranges of the series from which we derive its limit distribution. These nonparametric statistics endow the test with a number of desirable properties, the invariance to monotonic transformations of the series and the robustness to the presence of important parameter shifts. Moreover, the RUR test outperforms the power of standard unit-root tests on near-unit-root stationary time series; it is invariant with respect to the innovations distribution and asymptotically immune to noise. An extension of the RUR test, called the forward–backward range unit-root (FB-RUR) improves the check in the presence of additive outliers. Finally, we illustrate the performances of both range tests and their discrepancies with the Dickey–Fuller unit-root test on exchange rate series.